+++ collective/venue/behaviors.py
@@ -15,8 +15,8 @@ from zope.schema.interfaces import IContextAwareDefaultFactory
 def default_location(context):
     """Provide default location.
     """
-    default_venue = api.portal.get_registry_record('collective.venue.default_venue')  # noqa
-    return default_venue or ''
+    default = api.portal.get_registry_record('collective.venue.default_venue')
+    return default or ''
 
 
 @provider(IFormFieldProvider)
@@ -47,7 +47,71 @@ class ILocation(model.Schema):
             default=u'Location notes'),
         description=_(
             u'description_event_location_notes',
-            default=u'Additional Information for the Location.'),
+            default=u'One-time location or additional Information.'),
+        required=False,
+        default=None,
+    )
+
+    location_url = schema.URI(
+        title=_(
+            u'label_event_location_url',
+            default=u'location URL'),
+        description=_(
+            u'description_event_location_url',
+            default=u'One-time location URL.'),
+        required=False,
+        default=None,
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def default_organizer(context):
+    """Provide default organizer.
+    """
+    default = api.portal.get_registry_record('collective.venue.default_organizer')  # noqa
+    return default or ''
+
+
+@provider(IFormFieldProvider)
+class IOrganizer(model.Schema):
+
+    organizer_uid = schema.Choice(
+        title=_(u'label_event_organizer', default=u'Organizer'),
+        description=_(
+            u'description_event_organizer',
+            default=u'Select an organizer.'),
+        required=False,
+        missing_value='',
+        defaultFactory=default_organizer,
+        vocabulary='plone.app.vocabularies.Catalog',
+    )
+    form.widget(
+        'organizer_uid',
+        RelatedItemsFieldWidget,
+        pattern_options={
+            'selectableTypes': ['Venue'],
+            'basePath': get_base_path
+        }
+    )
+
+    organizer_notes = schema.Text(
+        title=_(
+            u'label_event_organizer_notes',
+            default=u'Organizer notes'),
+        description=_(
+            u'description_event_organizer_notes',
+            default=u'One-time organizer or additional Information.'),
+        required=False,
+        default=None,
+    )
+
+    organizer_url = schema.URI(
+        title=_(
+            u'label_event_organizer_url',
+            default=u'Organizer URL'),
+        description=_(
+            u'description_event_organizer_url',
+            default=u'One-time organizer URL.'),
         required=False,
         default=None,
     )